requirements.txt
asgiref==3.2.7
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-environ==0.4.5
djecrety==1.0.13
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
python-dotenv==0.14.0
python-http-client==3.3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.4.7
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
stripe==2.54.0
urllib3==1.25.8

dj-config-url~=0.1.1